refactor ~ simplify main 'lib' (repackage associated code)
